requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Cors==4.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
blinker==1.8.2
gunicorn==22.0.0
supabase
PyJWT
